requirements.txt
alembic==1.4.3
astroid==2.4.2
backcall==0.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
decorator==4.4.2
Deprecated==1.2.12
ecdsa==0.14.1
elasticsearch==7.9.1
Flask==1.1.2
flask-buzz==0.1.15
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
flask-praetorian==1.2.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.4
inflection==0.3.1
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.1
passlib==1.7.4
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py-buzz==1.0.3
pyasn1==0.4.8
pycparser==2.20
Pygments==2.7.1
PyJWT==2.0.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
pytzdata==2020.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
